% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_min_played.R
\name{highest_min_played}
\alias{highest_min_played}
\title{A nbahw5package Function}
\usage{
highest_min_played(year)
}
\arguments{
\item{year}{The year of interest you want to find the score from.}
}
\description{
This function allows you to find the player with the most minutes played in a specific year.
}
\examples{
highest_min_played()
}
\keyword{Max}
\keyword{minutes}
