% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michael_jordan_score.R
\name{michael_jordan_score}
\alias{michael_jordan_score}
\title{A nbahw5package Function}
\usage{
michael_jordan_score(year)
}
\arguments{
\item{year}{The year of interest you want to find the score from.}
}
\description{
This function allows you to find the score of Michael Jordan in a specific year.
}
\examples{
highest_mj_score()
}
\keyword{highest}
\keyword{jordan}
\keyword{michael}
\keyword{score}
