% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_corr_matrix.R
\name{nba_corr_matrix}
\alias{nba_corr_matrix}
\title{An nbahw5package function}
\usage{
nba_corr_matrix(year)
}
\arguments{
\item{year}{The season from which you want the correlation matrix.}
}
\description{
This function allows you to get the correlation matrix of continuous variables from a given season.
}
\examples{
continuous_correlation()
}
\keyword{continuous}
\keyword{correlation}
\keyword{matrix}
