% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_oldest_player.R
\name{nba_oldest_player}
\alias{nba_oldest_player}
\title{An nbahw5package function}
\usage{
nba_oldest_player(year)
}
\arguments{
\item{theyear}{the year of interest in which would you like to find the oldest player}
}
\description{
This function allows you to find the oldest player in the NBA data for that year
}
\examples{
nba_oldest_player()
}
\keyword{oldest}
\keyword{player}
